Even more warning fixes

+++ Controllers/CorsairPeripheralV2Controller/RGBController_CorsairV2Software.cpp
@@ -150,7 +150,7 @@ void RGBController_CorsairV2SW::SetupZones()
                     LOG_DEBUG("[%s] Created KB matrix with %d rows and %d columns containing %d keys",
                               controller->GetName().c_str(), new_kb.GetRowCount(), new_kb.GetColumnCount(), new_zone.leds_count);
 
-                    for(size_t led_idx = 0; led_idx < new_zone.leds_count; led_idx++)
+                    for(unsigned int led_idx = 0; led_idx < new_zone.leds_count; led_idx++)
                     {
                         led new_led;
