+++ Controllers/IonicoController/RGBController_Ionico.cpp
@@ -138,9 +138,9 @@ void RGBController_Ionico::SetupZones()
         zone zone_keyboard;
         zone_keyboard.name       = "Keyboard";
         zone_keyboard.type       = ZONE_TYPE_LINEAR;
-        zone_keyboard.leds_min   = leds.size();
-        zone_keyboard.leds_max   = leds.size();
-        zone_keyboard.leds_count = leds.size();
+        zone_keyboard.leds_min   = (unsigned int)leds.size();
+        zone_keyboard.leds_max   = (unsigned int)leds.size();
+        zone_keyboard.leds_count = (unsigned int)leds.size();
         zone_keyboard.matrix_map = nullptr;
         zones.emplace_back(zone_keyboard);
         for(size_t i = 0; i < leds.size(); ++i)
@@ -154,9 +154,9 @@ void RGBController_Ionico::SetupZones()
         zone zone_bar;
         zone_bar.name       = "Front Bar";
         zone_bar.type       = ZONE_TYPE_LINEAR;
-        zone_bar.leds_min   = leds.size();
-        zone_bar.leds_max   = leds.size();
-        zone_bar.leds_count = leds.size();
+        zone_bar.leds_min   = (unsigned int)leds.size();
+        zone_bar.leds_max   = (unsigned int)leds.size();
+        zone_bar.leds_count = (unsigned int)leds.size();
         zone_bar.matrix_map = nullptr;
         zones.emplace_back(zone_bar);
         for(size_t i = 0; i < leds.size(); ++i)
