+++ Controllers/LianLiController/LianLiUniHub_AL10Controller.cpp
@@ -754,14 +754,14 @@ void LianLiUniHub_AL10Controller::SendConfig(uint16_t wIndex, uint8_t  *config, 
     /*-------------------------------------*\
     | Send packet                           |
     \*-------------------------------------*/
-    size_t ret = libusb_control_transfer(handle,    /* dev_handle       */
-                                         0x40,      /* bmRequestType    */
-                                         0x80,      /* bRequest         */
-                                         0x00,      /* wValue           */
-                                         wIndex,    /* wIndex           */
-                                         config,    /* data             */
-                                         length,    /* wLength          */
-                                         1000);     /* timeout          */
+    size_t ret = libusb_control_transfer(handle,            /* dev_handle       */
+                                         0x40,              /* bmRequestType    */
+                                         0x80,              /* bRequest         */
+                                         0x00,              /* wValue           */
+                                         wIndex,            /* wIndex           */
+                                         config,            /* data             */
+                                         (uint16_t)length,  /* wLength          */
+                                         1000);             /* timeout          */
     std::this_thread::sleep_for(5ms);
     /*-------------------------------------*\
     | Check for communication error         |
